requirements.txt
streamlit>=1.38
plotly>=5.23
pandas>=2.2
numpy>=1.26
scipy>=1.14
xlrd>=2.0.1
pyarrow>=17.0
joblib>=1.4
tqdm>=4.66
psutil>=5.9
matplotlib>=3.9
seaborn>=0.13
watchdog>=4.0
scikit-learn==1.7.1
tqdm>=4.66.0
optuna==4.4.0
PyYAML==6.0.2
xgboost>=2.0.0
lightgbm>=4.3.0
catboost>=1.2.2
joblib>=1.3.2
openpyxl==3.1.5
category_encoders==2.8.1
requests==2.32.4
tensorboardX==2.6.4


